Added texture mapping and lighting, with image creation and play-back options

+++ Engine/Vec3.h
@@ -121,7 +121,7 @@ struct Vec3
 	}
 	inline float length()
 	{
-		return sqrt(lengthSqrd());
+		return (float)sqrt(lengthSqrd());
 	}
